# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  constants.PaymentMethod:
    enum:
    - Cash
    - BankTransfer
    - Credit
    - Company
    type: string
    x-enum-varnames:
    - PaymentMethodCash
    - PaymentMethodBankTransfer
    - PaymentMethodCredit
    - PaymentMethodCompany
...
  dto.ApproveExpenseRequest:
    properties:
      approval_comment:
        maxLength: 512
        type: string
    type: object
...
  dto.CreateExpenseRequest:
    properties:
      amount:
        type: number
      attachment_url:
        maxLength: 512
        type: string
      budget_item_id:
        type: integer
      description:
        maxLength: 512
        type: string
      expense_date:
        type: string
      invoice_no:
        maxLength: 128
        type: string
      payment_method:
        $ref: '#/definitions/constants.PaymentMethod'
      supplier_id:
        type: integer
    required:
    - amount
    - budget_item_id
    - expense_date
    - payment_method
    type: object
...
  dto.CreateRefundRequest:
    properties:
      amount:
        type: number
      reason:
        maxLength: 512
        type: string
      received_date:
        type: string
      voucher_no:
        maxLength: 128
        type: string
    required:
    - amount
    - received_date
    - voucher_no
    type: object
...
  dto.PayExpenseRequest:
    properties:
      payment_date:
        type: string
    type: object
  dto.RejectExpenseRequest:
    properties:
      approval_comment:
        maxLength: 512
        type: string
    required:
    - approval_comment
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:19306
info:
  contact:
    email: support@example.com
    name: API Support
  description: 装修预算与支出管理 API 服务，提供预算编制、支出审批、供应商管理与对账结算能力。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Renovation Budget API
  version: 1.0.0
paths:
...
  /expenses:
    get:
      parameters:
      - description: 审批状态
        in: query
        name: status
        type: string
      - description: 预算项ID
        in: query
        name: budget_item_id
        type: integer
      - description: 供应商ID
        in: query
        name: supplier_id
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 查询支出记录列表
      tags:
      - expenses
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建支出记录
      tags:
      - expenses
  /expenses/{id}:
    get:
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取支出记录详情
      tags:
      - expenses
  /expenses/{id}/approve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 审批意见
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApproveExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 审批通过支出
      tags:
      - expenses
  /expenses/{id}/pay:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 付款日期
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PayExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 确认付款
      tags:
      - expenses
  /expenses/{id}/refunds:
    get:
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 查询退款记录
      tags:
      - expenses
    post:
      consumes:
      - application/json
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 退款登记请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 登记退款
      tags:
      - expenses
  /expenses/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 驳回意见
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RejectExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 驳回支出
      tags:
      - expenses
  /expenses/{id}/submit:
    post:
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 提交支出
      tags:
      - expenses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"